+++ TeslaLogger/www/admin/password.php
@@ -418,7 +418,7 @@ if ($_REQUEST["id"] != -1)
 else
 {
 	$tinfo = get_text("INFO_FLEETAPI");
-	$tinfo=str_replace("{LINK1}", "<a href='https://developer.tesla.com/docs/fleet-api/support/announcements#2024-11-27-pay-per-use-pricing' target='_blank'>Tesla Pay per use pricing</a>", $tinfo);
+	$tinfo=str_replace("{LINK1}", "<a href='https://developer.tesla.com/docs/fleet-api/announcements#2024-11-27-pay-per-use-pricing' target='_blank'>Tesla Pay per use pricing</a>", $tinfo);
 	$tinfo=str_replace("{LINK2}", "<a href='https://digitalassets.tesla.com/tesla-contents/image/upload/Fleet-API-Agreement-EN.pdf' target='_blank'>Fleet API Agreement</a>", $tinfo);    
 ?>
 <div>
@@ -436,9 +436,9 @@ else
 			<th style='text-align:center;'><?php t("Free SUC"); ?></th>
 			<th style='text-align:center;'>Fleet API</th>
 			<th style='text-align:center;'>Virtual Key</th>
-			<th style='text-align:center;'>Access Type</th>
+			<!-- <th style='text-align:center;'>Access Type</th> -->
 			<th style='text-align:center;'>Signal Counter</th>
-			<th style='text-align:center;'>Subscription</th>
+			<th style='text-align:center;'><?php t("Subscription"); ?></th>
 			<th><?php t("Edit"); ?></th>
 		</tr>
 	</thead>
@@ -454,22 +454,30 @@ else
 			$vin = $v->{"vin"};
 			$tesla_carid = $v->{"tesla_carid"};
 			$access_type = $v->{"access_type"};
-			
+
+			$cartype = $v->{"car_type"};
+			$NeedSubscription = !($cartype === "models" || $cartype === "modelx" || $cartype === "models2");
 			
 			$freesuccheckbox = GetCheckbox($v->{"freesuc"});
-			$fleetAPICheckBox = GetCheckbox($v->{"fleetAPI"});
+			$fleetAPICheckBox = "";
+			
+			if ($v->{"fleetAPI"} == "0" && $NeedSubscription)
+				$fleetAPICheckBox = "<a href='password_fleet.php?id=$id&vin=$carVIN'>".get_text("FleetAPIRequired")." ⚠️</a>";
+			else
+				$fleetAPICheckBox = GetCheckbox($v->{"fleetAPI"});
+
 			$virtualKeyCheckBox = GetCheckbox($v->{"virtualkey"});
 
 			echo("	<tr>\r\n");
 			echo("		<td>$id</td>\r\n");
-			echo("		<td>$display_name</td>\r\n");
+			echo("		<td>$display_name <a href='changecarname.php?carid=$id'>&#9998</a></td>\r\n");
 			echo("		<td>$car</td>\r\n");
 			echo("		<td>$vin</td>\r\n");
 			echo("		<td>$tasker_token</td>\r\n");
 			echo("		<td style='text-align:center;'>$freesuccheckbox</td>\r\n");
 			echo("		<td style='text-align:center;'>$fleetAPICheckBox</td>\r\n");
 			echo("		<td style='text-align:center;'>$virtualKeyCheckBox</td>\r\n");
-			echo("		<td style='text-align:center;'>$access_type</td>\r\n");
+			// echo("		<td style='text-align:center;'>$access_type</td>\r\n");
 			
 			echo("		<td style='text-align:center;'>");
 			if ($v->{"fleetAPI"} == "1")
@@ -477,17 +485,17 @@ else
 			echo("</td>\r\n");
 
 			echo("		<td style='text-align:center;'>");
-			//if ($v->{"fleetAPI"} == "1")
+			if ($NeedSubscription)
 			{
 				$subscription = file_get_contents("https://teslalogger.de/stripe/subscription-check.php?vin=$vin");
 				
 				if (strpos($subscription, "current_period_end") > 0)
 				{
 					echo(GetCheckbox("1"));
-					echo("&nbsp;<a target='_blank' href='https://billing.stripe.com/p/login/8wMaGogxma56fGUdQQ'>Manage</a>");
+					echo("&nbsp;<a target='_blank' href='https://billing.stripe.com/p/login/8wMaGogxma56fGUdQQ'>". get_text("SubscribeManage") ."</a>");
 				}
 				else
-					echo("<a target='_blank' href='https://buy.stripe.com/9AQaHNdU33k29Vu144?client_reference_id=$vin'>Subscribe</a>");
+					echo("<a target='_blank' href='https://buy.stripe.com/9AQaHNdU33k29Vu144?client_reference_id=$vin'>⚠️ ". get_text("Subscribe") ."</a>");
 			}
 			echo("</td>\r\n");